x86/fpu: Correct the legacy state offset and size information
MXCSR is architecturally part of the SSE state. But, the kernel code
presumes it as part of the FP component. Adjust the offset and size for
these legacy states.

Notably, each legacy component area is not contiguous, unlike extended
components. Add a WARNING to be emitted when the location of those
legacy states is queried.

Fixes: ac73b27aea4e ("x86/fpu/xstate: Fix xstate_offsets, xstate_sizes for non-extended xstates")
Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org
---
Changes from v2:
* Replace pr_warn() with WARN_ON_ONCE() (Sean Christopherson).
* Massage the code comment (Sean Christopherson).