selftests/x86: Refactor altstack setup code
The sigaltstack setup code is almost the same across x86 tests. Most
of the test probably just needs a ready-to-use altstack instead of all
the setup code.

Refactor them to these helpers to simplify the test code:

- setup_sigaltstack()

  Allocate an altstack memory with a size more than the
  kernel-provided minimum. Then register the altstack via
  sigaltstack(2).

- cleanup_sigaltstack()

  Unregister the altstack and free up the memory.

Then, these two function calls are supposed to be enough for most
sigaltstack-needed test cases. But, when explicit sigaltstack() is
needed, another helper is available:

- init_sigaltstack()

  Do the same as setup_sigaltstack(), except for not invoking the
  sigaltstack syscall.

Finally, while here, ensure an altstack cleanup everywhere an altstack
is used.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Andy Lutomirski <luto@kernel.org>
Cc: linux-kselftest@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
---
The patch is based on another series that fixes the sigaltstack test:
  https://lore.kernel.org/all/20230330233520.21937-2-chang.seok.bae@intel.com/

Considerably that fix should go first before this cleanup rework.