x86/fpu/xstate: Simplify print_xstate_features()
The print_xstate_features() invokes print_xstate_feature() multiple
times in separate lines which can be replaced with a loop.

The print_xstate_feature() function already checks the feature
enabling status, and it is only called from within
print_xstate_features(). So the code can be relocated.

Move the code exactly to print_xstate_features() and wrap it with a
loop there.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org