x86/fpu: Remove the unnecessary fpstate_reset() invocation
fpstate_reset() initializes the fpstate container. But, when no state
has been established yet, calling this is not meaningful. In the
fpu__init_system() function, it is called immediately after the legacy
FPU size is determined. So, remove it as not causing any issues.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org
---

The change made by fpstate_reset() can be measured with this:

static struct fpstate fps;

void __init fpu__init_system(struct cpuinfo_x86 *c)
{
	memcpy(&fps, &current->thread.fpu.__fpstate, sizeof(struct fpstate));
	fpstate_reset(&current->thread.fpu);
	if (memcmp(&fps, &current->thread.fpu.__fpstate, sizeof(struct fpstate)))
		pr_info("fpstate has reason to be inited at %s\n", __func__);
	else
		pr_info("fpstate has no reason to be inited at %s\n", __func__);
	...
}

Although it looks to be a no-op, there might be some unknown
intention behind it. If there is a specific reason for invoking
fpstate_reset() at that point, this should be documented properly
instead of removing it.